Update 'from_spatial_coverages' to use MOC in the Rust store, remove dead code

+++ mocpy/abstract_moc.py
@@ -253,7 +253,7 @@ class AbstractMOC(serializer.IO):
         if args:
             store_indices = np.append(
                 [self._store_index, another_moc._store_index],
-                np.fromiter((arg._store_index for arg in args), dtype=np.int),
+                np.fromiter((arg._store_index for arg in args), dtype=np.uint),
             )
             index = mocpy.multi_intersection(store_indices)
         else:
@@ -280,7 +280,7 @@ class AbstractMOC(serializer.IO):
         if args:
             store_indices = np.append(
                 [self._store_index, another_moc._store_index],
-                np.fromiter((arg._store_index for arg in args), dtype=np.int),
+                np.fromiter((arg._store_index for arg in args), dtype=np.uint),
             )
             index = mocpy.multi_union(store_indices)
         else:
@@ -307,7 +307,7 @@ class AbstractMOC(serializer.IO):
         if args:
             store_indices = np.append(
                 [self._store_index, another_moc._store_index],
-                np.fromiter((arg._store_index for arg in args), dtype=np.int),
+                np.fromiter((arg._store_index for arg in args), dtype=np.uint),
             )
             index = mocpy.multi_symmetric_difference(store_indices)
         else:
@@ -336,7 +336,7 @@ class AbstractMOC(serializer.IO):
         if args:
             store_indices = np.append(
                 [another_moc._store_index],
-                np.fromiter((arg._store_index for arg in args), dtype=np.int),
+                np.fromiter((arg._store_index for arg in args), dtype=np.uint),
             )
             index_union = mocpy.multi_union(store_indices)
             index = mocpy.difference(self._store_index, index_union)
